requirements.txt
python-libtorrent